requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
openai==2.7.1
langchain==1.0.3
langchain-openai==1.0.2
langchain-community==0.4.1
chromadb==1.3.4
python-dotenv==1.2.1